app/requirements.txt
beautifulsoup4==4.12.3
celery==5.4.0
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
lxml_html_clean==0.3.1
newspaper4k==0.9.3.1
redis==5.2.0
requests==2.32.3
typing_extensions==4.12.2